requirements.txt
asn1crypto==1.1.0
blis==0.4.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cymem==2.0.2
Django==2.2.6
django-cors-headers==3.1.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
docx2txt==0.8
en-core-web-sm==2.2.0
idna==2.8
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
pdfminer.six==20181108
plac==0.9.6
preshed==3.0.2
pycparser==2.19
pycryptodome==3.9.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
spacy==2.2.1
sqlparse==0.3.0
srsly==0.1.0
thinc==7.1.1
tqdm==4.36.1
urllib3==1.25.6
wasabi==0.2.2
